Added IRequest interfaces to make it easier to raise requests in Mediator package
Core.Mediator (2.0.0):
-  Changed IMessageHandler.Priority type from ushort? to byte?
- Updated all components to use byte priority
- Added IRequest and IRequest<T> interfaces that requests need to implement before they can be raised

+++ Sels.Core.Mediator/Components/Event/DelegateEventListener.cs
@@ -20,7 +20,7 @@ namespace Sels.Core.Mediator.Event
 
         // Properties
         /// <inheritdoc/>
-        public ushort? Priority { get; set; }
+        public byte? Priority { get; set; }
 
         /// <inheritdoc cref="DelegateEventListener{TEvent}"/>
         /// <param name="asyncAction">The delegate that will be called to react to raised events</param>
@@ -43,7 +43,7 @@ namespace Sels.Core.Mediator.Event
 
         // Properties
         /// <inheritdoc/>
-        public ushort? Priority { get; set; }
+        public byte? Priority { get; set; }
 
         /// <inheritdoc cref="DelegateEventListener"/>
         /// <param name="asyncAction">The delegate that will be called to react to raised events</param>
